coordinator: extract operator-status accessors into status.rs
#985 Phase 7 — first impl-block extraction (vs prior field-only
phases). Move 16 read-only operator-status methods out of
coordinator/mod.rs into a sibling `impl super::Coordinator` block
in coordinator/status.rs:

- dynamic_neighbor_status, cos_no_owner_binding_drops_total
- recent_exceptions, recent_session_deltas, last_resolution
- slow_path_status, drain_session_deltas
- cos_statuses, filter_term_counters
- worker_heartbeats, worker_count, worker_runtime_snapshots
- identity_count, live_count, planned_counts, reconcile_debug

These are all pure-read methods called from the gRPC / HTTP status
surface; none mutate worker lifecycle, sessions, or HA reconciliation.
Cohesive cluster, easy to lift.

coordinator/mod.rs production LOC (excluding mod tests):
  2,191 → 2,028 (one step from the 2,000 threshold)
status.rs: 187 LOC.

Why now: post-Wave-6, all field-extraction phases hit diminishing
returns — manager extractions don't reduce LOC, only field count.
The remaining bulk is fn bodies. Splitting accessor methods into
sibling impl blocks lets us keep cutting LOC while preserving the
existing public Coordinator API.

build clean, 865 Rust tests pass.
